Modified Vertex an Tree
Made it so Vertex is hashable by it's id thus will be made unique and thus a Vertex added to a set will not be added if another instance of Vertex is already there with the same id

Made it posible to init Tree so that it doesn't have to start with a programme and can have many programmes added.
Also updated the regular expressions to match more instances like 'Any 178.1xx paper', etc.

git-svn-id: https://ist-linux1.massey.ac.nz/svn/jdhurfor/trunk/curriculum-information-representation@397 93a84c75-b7d8-402c-aa9c-cfdb53399b0d