Fixed up tests and made a pprint method for Tree
Made sure the vertex tests were testing for the right values.

Did something to programme_tree_test

Wrote a pprint method for Tree which sorts the papers into a list and prints them out indenting according to what level the paper is so a 300 level paper will be indented twice.