Fixed up tests and made a pprint method for Tree
Made sure the vertex tests were testing for the right values.

Did something to programme_tree_test

Wrote a pprint method for Tree which sorts the papers into a list and prints them out indenting according to what level the paper is so a 300 level paper will be indented twice.

git-svn-id: https://ist-linux1.massey.ac.nz/svn/jdhurfor/trunk/curriculum-information-representation@404 93a84c75-b7d8-402c-aa9c-cfdb53399b0d