Changed Tree.pprint parameter, rewrote the test to use this parameter, and altered build.xml
Tree.pprint now takes a output stream parameter which defaults to sys.stdout.  It can be of any output stream type which can be written to like a file.

The test uses this paprameter to pass to it a StringIO which I then use to get the result of the pprint and compare it with what I expect it to be.

build.xml has some minor changes done to the delete task.